+++ torchvision/csrc/cpu/image/writejpeg_cpu.h
@@ -2,8 +2,8 @@
 
 #include <torch/torch.h>
 
-C10_API torch::Tensor encodeJPEG(const torch::Tensor& data, int quality);
+C10_API torch::Tensor encodeJPEG(const torch::Tensor& data, int64_t quality);
 C10_API void writeJPEG(
     const torch::Tensor& data,
     const char* filename,
-    int quality);
+    int64_t quality);
